+++ test/map/map_builder_test.dart
@@ -85,6 +85,24 @@ void main() {
           {1: '1', 2: '2'});
     });
 
+    test('has addIterable method like Map.fromIterable', () {
+      expect(
+          (new MapBuilder<int, int>()..addIterable([1, 2, 3])).build().toMap(),
+          {1: 1, 2: 2, 3: 3});
+      expect(
+          (new MapBuilder<int, int>()
+                ..addIterable([1, 2, 3], key: (element) => element + 1))
+              .build()
+              .toMap(),
+          {2: 1, 3: 2, 4: 3});
+      expect(
+          (new MapBuilder<int, int>()
+                ..addIterable([1, 2, 3], value: (element) => element + 1))
+              .build()
+              .toMap(),
+          {1: 2, 2: 3, 3: 4});
+    });
+
     // Lazy copies.
 
     test('does not mutate BuiltMap following reuse of underlying Map', () {